+++ Include/Acidanthera/Library/OcAppleSecureBootLib.h
@@ -17,6 +17,18 @@
 #include <Protocol/AppleImg4Verification.h>
 
 /**
+  Bootstrap NVRAM and library values for secure booting.
+
+  @param[in] Model          Secure boot model (without ap suffix in lower-case).
+
+  @retval EFI_SUCCESS  On success.
+**/
+EFI_STATUS
+OcAppleSecureBootBootstrapValues (
+  IN CONST CHAR8  *Model
+  );
+
+/**
   Install and initialise the Apple Secure Boot protocol.
 
   @param[in] Reinstall          Replace any installed protocol.
@@ -26,7 +38,6 @@
 
   @returns Installed or located protocol.
   @retval NULL  There was an error locating or installing the protocol.
-
 **/
 APPLE_SECURE_BOOT_PROTOCOL *
 OcAppleSecureBootInstallProtocol (