+++ Library/OcAppleImg4Lib/OcAppleImg4Lib.c
@@ -17,11 +17,15 @@
 
 #include <Protocol/AppleImg4Verification.h>
 
+#include <Library/BaseLib.h>
 #include <Library/BaseMemoryLib.h>
 #include <Library/DebugLib.h>
 #include <Library/MemoryAllocationLib.h>
 #include <Library/OcAppleKeysLib.h>
 #include <Library/OcCryptoLib.h>
+#include <Library/OcAppleImg4Lib.h>
+#include <Library/OcMiscLib.h>
+#include <Library/UefiBootServicesTableLib.h>
 #include <Library/UefiRuntimeServicesTableLib.h>
 
 #include "libDER/oids.h"
@@ -30,6 +34,53 @@
 
 GLOBAL_REMOVE_IF_UNREFERENCED const UINT8 *DERImg4RootCertificate     = gAppleX86SecureBootRootCaCert;
 GLOBAL_REMOVE_IF_UNREFERENCED const UINTN *DERImg4RootCertificateSize = &gAppleX86SecureBootRootCaCertSize;
+
+typedef struct OC_SB_MODEL_DESC_ {
+  CONST CHAR8  *HardwareModel;
+  UINT32       BoardId;
+} OC_SB_MODEL_DESC;
+
+STATIC DERImg4Environment mEnvInfo;
+STATIC CONST CHAR8 *mModelDefault = "j215";
+STATIC OC_SB_MODEL_DESC mModelInformation[] = {
+  { "j137",  0x0A }, ///< iMacPro1,1
+  { "j680",  0x0B }, ///< MacBookPro15,1
+  { "j132",  0x0C }, ///< MacBookPro15,2
+  { "j140k", 0x17 }, ///< MacBookAir8,1
+  { "j140a", 0x37 }, ///< MacBookAir8,2
+  { "j174",  0x0E }, ///< Macmini8,1
+  { "j160",  0x0F }, ///< MacPro7,1
+  { "j780",  0x07 }, ///< MacBookPro15,3
+  { "j213",  0x18 }, ///< MacBookPro15,4
+  { "j152f", 0x3A }, ///< MacBookPro16,1
+  { "j214k", 0x3E }, ///< MacBookPro16,2
+  { "j230k", 0x3F }, ///< MacBookAir9,1
+  { "j223",  0x3B }, ///< MacBookPro16,3
+  { "j215",  0x38 }, ///< MacBookPro16,4
+  { "j185",  0x22 }, ///< iMac20,1 (?)
+  { "j185f", 0x23 }, ///< iMac20,2 (?)
+};
+
+STATIC BOOLEAN mHasDigestOverride;
+STATIC UINT8   mOriginalDigest[SHA384_DIGEST_SIZE];
+STATIC UINT8   mOverrideDigest[SHA384_DIGEST_SIZE];
+
+STATIC
+OC_SB_MODEL_DESC *
+InternalGetModelInfo (
+  IN CONST CHAR8  *Model
+  )
+{
+  UINTN  Index;
+
+  for (Index = 0; Index < ARRAY_SIZE (mModelInformation); ++Index) {
+    if (AsciiStrCmp (Model, mModelInformation[Index].HardwareModel) == 0) {
+      return &mModelInformation[Index];
+    }
+  }
+
+  return NULL;
+}
 
 bool
 DERImg4VerifySignature (
@@ -75,159 +126,28 @@ DERImg4VerifySignature (
            );
 }
 
-VOID
-InternalRetrieveHwInfo (
-  OUT DERImg4Environment  *Environment
+CONST CHAR8 *
+OcAppleImg4GetHardwareModel (
+  IN CONST CHAR8    *ModelRequest
   )
 {
-  EFI_STATUS Status;
-  UINTN      DataSize;
-  //
-  // FIXME: Retrieve these values from trusted storage and expose the variables.
-  //
-  ASSERT (Environment != NULL);
+  OC_SB_MODEL_DESC  *ModelInfo;
 
-  ZeroMem (Environment, sizeof (*Environment));
-
-  DataSize = sizeof (Environment->ecid);
-  Status = gRT->GetVariable (
-                  L"ApECID",
-                  &gAppleSecureBootVariableGuid,
-                  NULL,
-                  &DataSize,
-                  &Environment->ecid
-                  );
-  if (EFI_ERROR (Status) || DataSize != sizeof (Environment->ecid)) {
-    Environment->ecid = 0;
+  if (AsciiStrCmp (ModelRequest, OC_SB_MODEL_DEFAULT) == 0 || ModelRequest[0] == '\0') {
+    ModelRequest = mModelDefault;
   }
 
-  DataSize = sizeof (Environment->chipId);
-  Status = gRT->GetVariable (
-                  L"ApChipID",
-                  &gAppleSecureBootVariableGuid,
-                  NULL,
-                  &DataSize,
-                  &Environment->chipId
-                  );
-  if (EFI_ERROR (Status) || DataSize != sizeof (Environment->chipId)) {
-    Environment->chipId = 0;
+  ModelInfo = InternalGetModelInfo (ModelRequest);
+  if (ModelInfo != NULL) {
+    return ModelInfo->HardwareModel;
   }
 
-  DataSize = sizeof (Environment->boardId);
-  Status = gRT->GetVariable (
-                  L"ApBoardID",
-                  &gAppleSecureBootVariableGuid,
-                  NULL,
-                  &DataSize,
-                  &Environment->boardId
-                  );
-  if (EFI_ERROR (Status) || DataSize != sizeof (Environment->boardId)) {
-    Environment->boardId = 0;
-  }
-
-  Environment->certificateEpoch = 2;
-
-  DataSize = sizeof (Environment->securityDomain);
-  Status = gRT->GetVariable (
-                  L"ApSecurityDomain",
-                  &gAppleSecureBootVariableGuid,
-                  NULL,
-                  &DataSize,
-                  &Environment->securityDomain
-                  );
-  if (EFI_ERROR (Status) || DataSize != sizeof (Environment->securityDomain)) {
-    Environment->securityDomain = 1;
-  }
-
-  DataSize = sizeof (Environment->productionStatus);
-  Status = gRT->GetVariable (
-                  L"ApProductionStatus",
-                  &gAppleSecureBootVariableGuid,
-                  NULL,
-                  &DataSize,
-                  &Environment->productionStatus
-                  );
-  if (EFI_ERROR (Status)
-   || DataSize != sizeof (Environment->productionStatus)) {
-    Environment->productionStatus = TRUE;
-  }
-
-  DataSize = sizeof (Environment->securityMode);
-  Status = gRT->GetVariable (
-                  L"ApSecurityMode",
-                  &gAppleSecureBootVariableGuid,
-                  NULL,
-                  &DataSize,
-                  &Environment->securityMode
-                  );
-  if (EFI_ERROR (Status) || DataSize != sizeof (Environment->securityMode)) {
-    Environment->securityMode = TRUE;
-  }
-
-  DataSize = sizeof (Environment->effectiveProductionStatus);
-  Status = gRT->GetVariable (
-                  L"EffectiveProductionStatus",
-                  &gAppleSecureBootVariableGuid,
-                  NULL,
-                  &DataSize,
-                  &Environment->effectiveProductionStatus
-                  );
-  if (EFI_ERROR (Status)
-   || DataSize != sizeof (Environment->effectiveProductionStatus)) {
-    Environment->effectiveProductionStatus = TRUE;
-  }
-
-  DataSize = sizeof (Environment->effectiveSecurityMode);
-  Status = gRT->GetVariable (
-                  L"EffectiveSecurityMode",
-                  &gAppleSecureBootVariableGuid,
-                  NULL,
-                  &DataSize,
-                  &Environment->effectiveSecurityMode
-                  );
-  if (EFI_ERROR (Status)
-   || DataSize != sizeof (Environment->effectiveSecurityMode)) {
-    Environment->effectiveSecurityMode = TRUE;
-  }
-
-  DataSize = sizeof (Environment->internalUseOnlyUnit);
-  Status = gRT->GetVariable (
-                  L"InternalUseOnlyUnit",
-                  &gAppleSecureBootVariableGuid,
-                  NULL,
-                  &DataSize,
-                  &Environment->internalUseOnlyUnit
-                  );
-  if (EFI_ERROR (Status)
-   || DataSize != sizeof (Environment->internalUseOnlyUnit)) {
-    Environment->internalUseOnlyUnit = FALSE;
-  }
-
-  // CHANGE: HardwareModel is unused.
+  return NULL;
 }
 
-/**
-  Verify the signature of ImageBuffer against Type of its IMG4 Manifest.
-
-  @param[in]  This            The pointer to the current protocol instance.
-  @param[in]  ObjType         The IMG4 object type to validate against.
-  @param[in]  ImageBuffer     The buffer to validate.
-  @param[in]  ImageSize       The size, in bytes, of ImageBuffer.
-  @param[in]  SbMode          The requested IMG4 Secure Boot mode.
-  @param[in]  ManifestBuffer  The buffer of the IMG4 Manifest.
-  @param[in]  ManifestSize    The size, in bytes, of ManifestBuffer.
-  @param[out] HashDigest      On output, a pointer to ImageBuffer's digest.
-  @param[out] DigestSize      On output, the size, in bytes, of *HashDigest.
-
-  @retval EFI_SUCCESS             ImageBuffer is correctly signed.
-  @retval EFI_INVALID_PARAMETER   One or more required parameters are NULL.
-  @retval EFI_OUT_OF_RESOURCES    Not enough resources are available.
-  @retval EFI_SECURITY_VIOLATION  ImageBuffer's signature is invalid.
-
-**/
 EFI_STATUS
 EFIAPI
-AppleImg4Verify (
+OcAppleImg4Verify (
   IN  APPLE_IMG4_VERIFICATION_PROTOCOL  *This,
   IN  UINT32                            ObjType,
   IN  CONST VOID                        *ImageBuffer,
@@ -241,8 +161,8 @@ AppleImg4Verify (
 {
   DERReturn           DerResult;
   INTN                CmpResult;
+  UINT8               Digest[SHA384_DIGEST_SIZE];
 
-  DERImg4Environment  EnvInfo;
   DERImg4ManifestInfo ManInfo;
 
   if ((ImageBuffer    == NULL || ImageSize    == 0)
@@ -263,48 +183,83 @@ AppleImg4Verify (
   if (DerResult != DR_Success) {
     return EFI_SECURITY_VIOLATION;
   }
+
+  CmpResult = -1;
+
+  //
+  // Provide a route to accept our modified kernel as long as we can trust it is really it.
+  //
+  if (mHasDigestOverride) {
+    DEBUG ((
+      DEBUG_INFO,
+      "OCI4: Trying override %u vs %u for %02X%02X%02X%02X\n",
+      ManInfo.imageDigestSize,
+      SHA384_DIGEST_SIZE,
+      ManInfo.imageDigest[0],
+      ManInfo.imageDigest[1],
+      ManInfo.imageDigest[2],
+      ManInfo.imageDigest[3]
+      ));
+    if (ManInfo.imageDigestSize == SHA384_DIGEST_SIZE) {
+      Sha384 (Digest, ImageBuffer, ImageSize);
+      if (CompareMem (Digest, mOverrideDigest, sizeof (mOverrideDigest)) == 0
+        && CompareMem (mOriginalDigest, ManInfo.imageDigest, sizeof (mOriginalDigest)) == 0) {
+        DEBUG ((
+          DEBUG_INFO,
+          "OCI4: Digest matched %02X%02X%02X%02X, accepting and disabling\n",
+          mOriginalDigest[0],
+          mOriginalDigest[1],
+          mOriginalDigest[2],
+          mOriginalDigest[3]
+          ));
+        CmpResult = 0;
+        mHasDigestOverride = FALSE;
+      }
+    }
+  }
+
   //
   // As ManInfo.imageDigest is a buffer of static size, the bounds check to
   // retrieve it acts as implicit sanitizing of ManInfo.imageDigestSize which
   // can be considered trusted at this point.
   //
-  CmpResult = SigVerifyShaHashBySize (
-                ImageBuffer,
-                ImageSize,
-                ManInfo.imageDigest,
-                ManInfo.imageDigestSize
-                );
+  if (CmpResult != 0) {
+    CmpResult = SigVerifyShaHashBySize (
+      ImageBuffer,
+      ImageSize,
+      ManInfo.imageDigest,
+      ManInfo.imageDigestSize
+      );
+  }
   if (CmpResult != 0) {
     return EFI_SECURITY_VIOLATION;
   }
 
-  InternalRetrieveHwInfo (&EnvInfo);
-
   if (SbMode == AppleImg4SbModeMedium) {
     if (ManInfo.hasEcid
-     || !ManInfo.hasXugs || ManInfo.environment.xugs != EnvInfo.xugs
+     || !ManInfo.hasXugs || ManInfo.environment.xugs != mEnvInfo.xugs
      || (ManInfo.environment.internalUseOnlyUnit
-      && ManInfo.environment.internalUseOnlyUnit != EnvInfo.internalUseOnlyUnit)) {
+      && ManInfo.environment.internalUseOnlyUnit != mEnvInfo.internalUseOnlyUnit)) {
       return EFI_SECURITY_VIOLATION;
     }
   } else if (SbMode == AppleImg4SbModeFull) {
-    if (!ManInfo.hasEcid || ManInfo.environment.ecid != EnvInfo.ecid
+    if (!ManInfo.hasEcid || ManInfo.environment.ecid != mEnvInfo.ecid
      || ManInfo.hasXugs
      || ManInfo.environment.internalUseOnlyUnit) {
       return EFI_SECURITY_VIOLATION;
     }
   }
 
-  if ((ManInfo.environment.boardId                    != EnvInfo.boardId)
-   || (ManInfo.environment.chipId                     != EnvInfo.chipId)
-   || (ManInfo.environment.certificateEpoch            < EnvInfo.certificateEpoch)
-   || (ManInfo.environment.productionStatus           != EnvInfo.productionStatus && !ManInfo.environment.productionStatus)
-   || (ManInfo.environment.securityMode               != EnvInfo.securityMode && !ManInfo.environment.securityMode)
-   || (ManInfo.environment.securityDomain             != EnvInfo.securityDomain)
+  if ((ManInfo.environment.boardId                    != mEnvInfo.boardId)
+   || (ManInfo.environment.chipId                     != mEnvInfo.chipId)
+   || (ManInfo.environment.certificateEpoch            < mEnvInfo.certificateEpoch)
+   || (ManInfo.environment.productionStatus           != mEnvInfo.productionStatus && !ManInfo.environment.productionStatus)
+   || (ManInfo.environment.securityMode               != mEnvInfo.securityMode && !ManInfo.environment.securityMode)
+   || (ManInfo.environment.securityDomain             != mEnvInfo.securityDomain)
    || (ManInfo.hasEffectiveProductionStatus
-    && (ManInfo.environment.effectiveProductionStatus != EnvInfo.effectiveProductionStatus && !ManInfo.environment.effectiveProductionStatus))
+    && (ManInfo.environment.effectiveProductionStatus != mEnvInfo.effectiveProductionStatus && !ManInfo.environment.effectiveProductionStatus))
    || (ManInfo.hasEffectiveSecurityMode
-    && (ManInfo.environment.effectiveSecurityMode     != EnvInfo.effectiveSecurityMode && !ManInfo.environment.effectiveSecurityMode))
+    && (ManInfo.environment.effectiveSecurityMode     != mEnvInfo.effectiveSecurityMode && !ManInfo.environment.effectiveSecurityMode))
     ) {
     return EFI_SECURITY_VIOLATION;
   }
@@ -319,4 +274,198 @@ AppleImg4Verify (
   }
 
   return EFI_SUCCESS;
+}
+
+VOID
+OcAppleImg4RegisterOverride (
+  IN CONST UINT8  *OriginalDigest,
+  IN CONST UINT8  *Image,
+  IN UINT32       ImageSize
+  )
+{
+  ASSERT (OriginalDigest != NULL);
+  ASSERT (Image != NULL);
+  ASSERT (ImageSize > 0);
+
+  mHasDigestOverride = TRUE;
+  CopyMem (mOriginalDigest, OriginalDigest, sizeof (mOriginalDigest));
+  Sha384 (mOverrideDigest, Image, ImageSize);
+}
+
+EFI_STATUS
+OcAppleImg4BootstrapValues (
+  IN CONST CHAR8   *Model,
+  IN UINT64        Ecid  OPTIONAL
+  )
+{
+  EFI_STATUS        Status;
+  OC_SB_MODEL_DESC  *SbInfo;
+
+  ASSERT (Model != NULL);
+  SbInfo = InternalGetModelInfo (Model);
+  ASSERT (SbInfo != NULL); ///< Checked by calling OcAppleImg4GetHardwareModel.
+
+  //
+  // Predefine most of the values as they are common for all the models.
+  //
+  mEnvInfo.ecid                       = Ecid;
+  mEnvInfo.boardId                    = SbInfo->BoardId;
+  mEnvInfo.chipId                     = 0x8012;
+  mEnvInfo.certificateEpoch           = 2;
+  mEnvInfo.securityDomain             = 1;
+  mEnvInfo.productionStatus           = TRUE;
+  mEnvInfo.securityMode               = 1;
+  mEnvInfo.effectiveProductionStatus  = TRUE;
+  mEnvInfo.effectiveSecurityMode      = 1;
+  mEnvInfo.internalUseOnlyUnit        = FALSE;
+  mEnvInfo.xugs                       = 1;
+
+  //
+  // Expose all the variables via NVRAM.
+  //
+  Status = gRT->SetVariable (
+    L"ApECID",
+    &gAppleSecureBootVariableGuid,
+    EFI_VARIABLE_RUNTIME_ACCESS | EFI_VARIABLE_BOOTSERVICE_ACCESS,
+    sizeof (mEnvInfo.ecid),
+    &mEnvInfo.ecid
+    );
+  if (EFI_ERROR (Status)) {
+    return Status;
+  }
+
+  Status = gRT->SetVariable (
+    L"ApChipID",
+    &gAppleSecureBootVariableGuid,
+    EFI_VARIABLE_RUNTIME_ACCESS | EFI_VARIABLE_BOOTSERVICE_ACCESS,
+    sizeof (mEnvInfo.chipId),
+    &mEnvInfo.chipId
+    );
+  if (EFI_ERROR (Status)) {
+    return Status;
+  }
+
+  Status = gRT->SetVariable (
+    L"ApBoardID",
+    &gAppleSecureBootVariableGuid,
+    EFI_VARIABLE_RUNTIME_ACCESS | EFI_VARIABLE_BOOTSERVICE_ACCESS,
+    sizeof (mEnvInfo.boardId),
+    &mEnvInfo.boardId
+    );
+  if (EFI_ERROR (Status)) {
+    return Status;
+  }
+
+  Status = gRT->SetVariable (
+    L"ApSecurityDomain",
+    &gAppleSecureBootVariableGuid,
+    EFI_VARIABLE_RUNTIME_ACCESS | EFI_VARIABLE_BOOTSERVICE_ACCESS,
+    sizeof (mEnvInfo.securityDomain),
+    &mEnvInfo.securityDomain
+    );
+  if (EFI_ERROR (Status)) {
+    return Status;
+  }
+
+  Status = gRT->SetVariable (
+    L"ApProductionStatus",
+    &gAppleSecureBootVariableGuid,
+    EFI_VARIABLE_RUNTIME_ACCESS | EFI_VARIABLE_BOOTSERVICE_ACCESS,
+    sizeof (mEnvInfo.productionStatus),
+    &mEnvInfo.productionStatus
+    );
+  if (EFI_ERROR (Status)) {
+    return Status;
+  }
+
+  Status = gRT->SetVariable (
+    L"ApSecurityMode",
+    &gAppleSecureBootVariableGuid,
+    EFI_VARIABLE_RUNTIME_ACCESS | EFI_VARIABLE_BOOTSERVICE_ACCESS,
+    sizeof (mEnvInfo.securityMode),
+    &mEnvInfo.securityMode
+    );
+  if (EFI_ERROR (Status)) {
+    return Status;
+  }
+
+  Status = gRT->SetVariable (
+    L"EffectiveProductionStatus",
+    &gAppleSecureBootVariableGuid,
+    EFI_VARIABLE_RUNTIME_ACCESS | EFI_VARIABLE_BOOTSERVICE_ACCESS,
+    sizeof (mEnvInfo.effectiveProductionStatus),
+    &mEnvInfo.effectiveProductionStatus
+    );
+  if (EFI_ERROR (Status)) {
+    return Status;
+  }
+
+  Status = gRT->SetVariable (
+    L"EffectiveSecurityMode",
+    &gAppleSecureBootVariableGuid,
+    EFI_VARIABLE_RUNTIME_ACCESS | EFI_VARIABLE_BOOTSERVICE_ACCESS,
+    sizeof (mEnvInfo.effectiveSecurityMode),
+    &mEnvInfo.effectiveSecurityMode
+    );
+  if (EFI_ERROR (Status)) {
+    return Status;
+  }
+
+  Status = gRT->SetVariable (
+    L"InternalUseOnlyUnit",
+    &gAppleSecureBootVariableGuid,
+    EFI_VARIABLE_RUNTIME_ACCESS | EFI_VARIABLE_BOOTSERVICE_ACCESS,
+    sizeof (mEnvInfo.internalUseOnlyUnit),
+    &mEnvInfo.internalUseOnlyUnit
+    );
+  if (EFI_ERROR (Status)) {
+    return Status;
+  }
+
+  return EFI_SUCCESS;
+}
+
+APPLE_IMG4_VERIFICATION_PROTOCOL *
+OcAppleImg4VerificationInstallProtocol (
+  IN BOOLEAN  Reinstall
+  )
+{
+  STATIC APPLE_IMG4_VERIFICATION_PROTOCOL Img4Verification = {
+    APPLE_IMG4_VERIFICATION_PROTOCOL_REVISION,
+    OcAppleImg4Verify
+  };
+
+  EFI_STATUS                       Status;
+  APPLE_IMG4_VERIFICATION_PROTOCOL *Protocol;
+  EFI_HANDLE                       Handle;
+
+  if (Reinstall) {
+    Status = OcUninstallAllProtocolInstances (&gAppleImg4VerificationProtocolGuid);
+    if (EFI_ERROR (Status)) {
+      DEBUG ((DEBUG_ERROR, "OCI4: Uninstall failed: %r\n", Status));
+      return NULL;
+    }
+  } else {
+    Status = gBS->LocateProtocol (
+      &gAppleImg4VerificationProtocolGuid,
+      NULL,
+      (VOID **)&Protocol
+      );
+    if (!EFI_ERROR (Status)) {
+      return Protocol;
+    }
+  }
+
+  Handle = NULL;
+  Status = gBS->InstallMultipleProtocolInterfaces (
+    &Handle,
+    &gAppleImg4VerificationProtocolGuid,
+    (VOID **)&Img4Verification,
+    NULL
+    );
+  if (EFI_ERROR (Status)) {
+    return NULL;
+  }
+
+  return &Img4Verification;
 }