+++ Library/OcAppleSecureBootLib/OcAppleSecureBootLib.c
@@ -20,6 +20,7 @@ WITHOUT WARRANTIES OR REPRESENTATIONS OF ANY KIND, EITHER EXPRESS OR IMPLIED.
 #include <Protocol/AppleImg4Verification.h>
 #include <Protocol/SimpleFileSystem.h>
 
+#include <Library/BaseLib.h>
 #include <Library/BaseMemoryLib.h>
 #include <Library/DebugLib.h>
 #include <Library/MemoryAllocationLib.h>
@@ -37,6 +38,7 @@ STATIC BOOLEAN mSbAvailable = TRUE;
 STATIC UINT8   mSbPolicy             = AppleImg4SbModeMedium;
 STATIC UINT8   mSbWindowsPolicy      = 1;
 STATIC BOOLEAN mSbWindowsPolicyValid = TRUE;
+STATIC CHAR8   mSbHardwareModel[16];
 
 STATIC
 UINT8
@@ -80,35 +82,35 @@ InternalImg4GetFailureReason (
   return Reason;
 }
 
-STATIC
-BOOLEAN
-InternalGetHardwareModel (
-  OUT CHAR8  Model[16]
+EFI_STATUS
+OcAppleSecureBootBootstrapValues (
+  IN CONST CHAR8  *Model
   )
-{ 
-  //
-  // FIXME: Retrieve this value from trusted storage and expose the variable.
-  //
-  EFI_STATUS Status;
-  UINTN      DataSize;
-  CHAR8      HwModel[16];
+{
+  EFI_STATUS  Status;
 
   ASSERT (Model != NULL);
 
-  DataSize = sizeof (HwModel);
-  Status = gRT->GetVariable (
-                  L"HardwareModel",
-                  &gAppleSecureBootVariableGuid,
-                  NULL,
-                  &DataSize,
-                  &HwModel
-                  );
-  if (EFI_ERROR (Status) || HwModel[DataSize - 1] != '\0') {
-    return FALSE;
+  Status = OcAsciiSafeSPrint (
+    mSbHardwareModel,
+    sizeof (mSbHardwareModel),
+    "%aap",
+    Model
+    );
+
+  if (EFI_ERROR (Status)) {
+    return Status;
   }
 
-  CopyMem (Model, HwModel, DataSize);
-  return TRUE;
+  Status = gRT->SetVariable (
+    L"HardwareModel",
+    &gAppleSecureBootVariableGuid,
+    EFI_VARIABLE_RUNTIME_ACCESS | EFI_VARIABLE_BOOTSERVICE_ACCESS,
+    AsciiStrSize (mSbHardwareModel),
+    mSbHardwareModel
+    );
+
+  return Status;
 }
 
 STATIC
@@ -579,11 +581,10 @@ InternalVerifyImg4ByPathWorker (
   VOID                            *ManifestBuffer;
   UINT32                          ManifestSize;
 
-  CHAR8                           HwModel[16];
   UINT64                          Ecid;
 
   STATIC CONST UINTN ManifestSuffixMaxSize =
-    (((ARRAY_SIZE (HwModel) - 1 + (2 * sizeof (Ecid))) * sizeof (CHAR16))
+    (((ARRAY_SIZE (mSbHardwareModel) - 1 + (2 * sizeof (Ecid))) * sizeof (CHAR16))
       + L_STR_SIZE_NT (L"...im4m"));
 
   Status = gBS->LocateDevicePath (
@@ -640,7 +641,7 @@ InternalVerifyImg4ByPathWorker (
     return EFI_NO_MEDIA;
   }
 
-  Result = InternalGetHardwareModel (HwModel);
+  Result = mSbHardwareModel[0] != '\0';
   if (Result) {
     ManifestSuffix = &Path[(ImagePathSize / sizeof (*Path)) - 1];
     if (SbPolicy == AppleImg4SbModeMedium) {
@@ -648,7 +649,7 @@ InternalVerifyImg4ByPathWorker (
         ManifestSuffix,
         ManifestSuffixMaxSize,
         L".%a.im4m",
-        HwModel
+        mSbHardwareModel
         );
     } else if (SbPolicy == AppleImg4SbModeFull) {
       Result = InternalGetApEcid (&Ecid);
@@ -657,7 +658,7 @@ InternalVerifyImg4ByPathWorker (
           ManifestSuffix,
           ManifestSuffixMaxSize,
           L".%a.%LX.im4m",
-          HwModel,
+          mSbHardwareModel,
           Ecid
           );
       }