+++ Changelog.md
@@ -8,6 +8,8 @@ OpenCore Changelog
 - Added kext injection support without kernel caches
 - Added Comet Lake-LP HDA device code
 - Fixed OS boot selection on SATA controllers with legacy OPROMs
+- Fixed RSDP ACPI table checksum recalculation
+- Added immutablekernel loading support for 10.13+
 
 #### v0.6.0
 - Fixed sound corruption with AudioDxe